Version 1.2
Option to open a web browser after the paste is posted to the pastebin site.  See the paster-config.vim file for details.  Set the g:PASTER_BROWSER_COMMAND variable in .vimrc with the command for opening the browser.

Michael Jansen provided a pastey bug fix and ideas for creating the browser interface.